Move ScriptingDocumentationWidget to picard.ui.widgets
